Split into main file and module Fixes #3
