+++ password_checker/__init__.py
@@ -1,0 +1,3 @@
+from password_checker.app import hash_password, check_password
+
+__all__ = ["hash_password", "check_password"]
